changed billing dto double to bigdecimal

+++ src/main/java/com/example/adminDashboardProject/dto/BillingDTO.java
@@ -1,6 +1,9 @@
 package com.example.adminDashboardProject.dto;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 
@@ -12,5 +15,5 @@ public class BillingDTO {
     private String variantUniqueId; // Business SKU (useful for barcode scanners)
     private Integer quantity;       // How many were bought
     private Long productId;     // Useful for front-end display
-    private Double totalPrice;
+    private BigDecimal totalPrice;
 }
